+++ aiter/ops/triton/_triton_kernels/attention/fav3_sage_attention_mxfp4.py
@@ -7,9 +7,6 @@ import aiter
 
 
 from aiter.ops.triton.utils._triton.pid_preprocessing import pid_grid_3d
-from aiter.ops.triton._triton_kernels.attention.fav3_sage_attention import (
-    map_dims,
-)
 
 
 @triton.jit
