+++ aiter/ops/triton/attention/fav3_sage.py
@@ -8,12 +8,22 @@ import aiter
 import triton
 from aiter.ops.triton._triton_kernels.attention.fav3_sage_attention import (
     sage_fwd,
-    map_dims,
-    triton_bmm_pool_sim_simmean,
-    triton_fill_block_map_kernel,
-    triton_fill_causal_mask,
 )
-from aiter.ops.triton.attention.utils import block_attn_mask_to_ragged_lut
+
+# The ``map_dims`` layout helper lives in attention/utils.py; the Sparge / VFA
+# block-sparse mask, ragged-LUT, and m_init preparation helpers live in
+# attention/block_sparse.py. They are re-exported here so existing
+# ``from aiter.ops.triton.attention.fav3_sage import ...`` call sites keep working.
+from aiter.ops.triton.attention.utils import map_dims  # noqa: F401
+from aiter.ops.triton.attention.block_sparse import (  # noqa: F401
+    block_attn_mask_to_ragged_lut,
+    fill_block_map_triton,
+    fill_causal_mask_triton,
+    get_block_map_meansim,
+    block_attn_mask_to_ragged_lut_topn_front,
+    build_attention_lut,
+    compute_m_proxy_topn,
+)
 from aiter.ops.triton.quant.sage_attention_quant_wrappers import sage_quant
 
 from aiter.ops.triton.utils._triton import arch_info
@@ -374,6 +384,7 @@ def fav3_sage_func(
     use_block_sparse: bool = False,
     freeze_softmax_max_count: int = -1,
     sparge_load_balancing: bool = False,
+    m_init: Optional[torch.Tensor] = None,
 ):
     """
     SageAttention v1.
@@ -406,6 +417,16 @@ def fav3_sage_func(
                 ``acc += p @ v`` with ``m`` held fixed (VFA-style). ``-1``
                 (default) disables freezing and keeps the exact online softmax.
                 Only takes effect on the block-sparse path.
+        m_init: optional precomputed per-row running-max estimate (VFA), fp32
+                ``[batch, nheads_q, num_q_blocks, BLKQ]``. When provided the
+                kernel skips the online-softmax rowmax reduction and the acc
+                rescale entirely (frozen max from the first block), computing
+                ``p = exp2(qk - m_init)``, ``l += rowsum(p)`` and
+                ``acc += p @ v``. Works for the dense non-causal path and the
+                block-sparse path. Mutually exclusive with causal/sliding-window
+                masking and with ``freeze_softmax_max_count``. See
+                :func:`compute_m_proxy_topn` for the estimator. ``None``
+                (default) runs the exact online softmax.
 
     Returns:
         out: Output tensor [batch, seqlen, num_q_heads, head_dim] or [batch, num_q_heads, seqlen, head_dim] (FP32)
@@ -497,6 +518,32 @@ def fav3_sage_func(
             "use_block_sparse=True; leave it at -1 (disabled) otherwise."
         )
 
+    # VFA: a precomputed frozen running-max estimate. Incompatible with the
+    # masking paths (causal/sliding window) and with the warm-up freeze, which
+    # derive their own running max.
+    use_precomputed_max = m_init is not None
+    if use_precomputed_max:
+        if causal or use_sliding_window:
+            raise NotImplementedError(
+                "m_init (VFA precomputed max) does not support causal or "
+                "sliding-window masking; require causal=False and "
+                "window_size=(-1, -1)."
+            )
+        if freeze_softmax_max_count >= 0:
+            raise ValueError(
+                "m_init (VFA precomputed max) and freeze_softmax_max_count are "
+                "mutually exclusive; pass only one."
+            )
+        assert m_init.dtype == torch.float32, "m_init must be fp32"
+        assert m_init.shape == (batch, nheads_q, num_q_blocks, BLKQ), (
+            f"m_init shape {tuple(m_init.shape)} does not match expected "
+            f"{(batch, nheads_q, num_q_blocks, BLKQ)}"
+        )
+        stride_mz, stride_mh, stride_mblk, stride_mr = m_init.stride()
+    else:
+        m_init = torch.zeros(1, dtype=torch.float32, device=q.device)
+        stride_mz = stride_mh = stride_mblk = stride_mr = 0
+
     if use_block_sparse:
         if kv_block_indices is None or lut_start is None or lut_count is None:
             raise ValueError(
@@ -566,6 +613,11 @@ def fav3_sage_func(
         stride_ksblk,
         stride_vsz,
         stride_vsh,
+        m_init,
+        stride_mz,
+        stride_mh,
+        stride_mblk,
+        stride_mr,
         softmax_lse,
         out,
         None,
@@ -632,6 +684,7 @@ def fav3_sage_func(
         USE_SEQUSED=False,
         USE_BLOCK_SPARSE=use_block_sparse,
         FREEZE_SOFTMAX_MAX_COUNT=freeze_softmax_max_count,
+        USE_PRECOMPUTED_MAX=use_precomputed_max,
         TILE_SCHEDULE=tile_schedule,
         TILE_COUNTER=tile_counter,
         NUM_TILES=n_tiles,
@@ -645,435 +698,82 @@ def fav3_sage_func(
         return out, None
 
 
-def get_pool_sim_triton_simmean(
-    x: torch.Tensor,
-    block_size: int,
-    simthreshd1: torch.Tensor,
-    attention_scored_only: bool = False,
-) -> Tuple[torch.Tensor, Optional[torch.Tensor]]:
-    """Mean-pool each block and flag internally-similar blocks.
-
-    Args:
-        x: ``(B, H, N, D)`` tensor.
-        block_size: number of tokens per block.
-        simthreshd1: ``(H,)`` per-head similarity threshold.
-        attention_scored_only: when ``True``, skip the intra-block similarity
-            test entirely and return ``None`` for ``sim_blocks``.
-
-    Steps:
-        1. Pool (mean) within each block.
-        2. Compute the mean pairwise cosine similarity within each block.
-        3. Flag blocks whose mean self-similarity exceeds ``simthreshd1``.
-
-    Note how the 3rd dimension ``N`` is reduced to ``nblock = N // block_size``;
-    this keeps the downstream block-attention proxy at ``O(nblock^2)`` instead
-    of the full ``O(N^2)``.
-
-    Returns:
-        pool: ``(B, H, nblock, D)`` tensor.
-        sim_blocks: ``(B, H, nblock)`` bool tensor, or ``None`` when
-            ``attention_scored_only`` is set.
-    """
-    x = x.contiguous()
-    B, H, N, D = x.shape
-    nblock = (N + block_size - 1) // block_size  # Number of blocks per feature map
-    pool = torch.empty((B, H, nblock, D), device=x.device, dtype=x.dtype)
-    if attention_scored_only:
-        sim_blocks = None
-        # The kernel needs a valid pointer; pass `pool` as an unused placeholder.
-        sim_arg = pool
-    else:
-        sim_blocks = torch.empty((B, H, nblock), device=x.device, dtype=torch.bool)
-        sim_arg = sim_blocks
-    grid = (B, H, nblock)
-    triton_bmm_pool_sim_simmean[grid](
-        x, pool, sim_arg, simthreshd1, N=N, D=D, BS=block_size, SKIP_SIM=attention_scored_only
-    )
-    return pool, sim_blocks
-
-
-def fill_block_map_triton(
-    final_map: torch.Tensor,
-    num_to_select: torch.Tensor,
-    sorted_indices: torch.Tensor,
-) -> torch.Tensor:
-    """Scatter the top-``num_to_select`` ranked K blocks per (B, H, Q) into ``final_map``."""
-    final_map = final_map.contiguous()
-    num_to_select = num_to_select.contiguous()
-    sorted_indices = sorted_indices.contiguous()
-    B, H, Q, K = final_map.shape
-    grid = (B, H, Q)
-    triton_fill_block_map_kernel[grid](final_map, num_to_select, sorted_indices, K)
-    return final_map
-
-
-def fill_causal_mask_triton(mask: torch.Tensor, BqdivBk: float) -> torch.Tensor:
-    """Fill a 2-D ``(nq, nk)`` block-level causal mask for a Q/K block-size ratio."""
-    assert mask.dim() == 2
-    triton_fill_causal_mask[mask.shape](mask, BqdivBk)
-    return mask
-
-
-def get_block_map_meansim(
+def fav3_sage_vfa_wrapper_func(
     q: torch.Tensor,
     k: torch.Tensor,
-    is_causal: bool = False,
-    BLKQ: int = 64,
-    BLKK: int = 64,
-    simthreshd1: float = 0.1,
-    cdfthreshd: float = 0.9,
-    attention_sink: bool = False,
-    attention_scored_only: bool = False,
-) -> Tuple[Optional[torch.Tensor], torch.Tensor]:
-    """Build a SpargeAttn block-sparse mask via the mean-similarity proxy.
+    v: torch.Tensor,
+    softmax_scale: Optional[float] = None,
+    return_lse: bool = False,
+    layout: str = "bshd",
+    config: Optional[dict] = None,
+    n_sample_blocks: int = 16,
+    block_lut: Optional[Tuple[torch.Tensor, torch.Tensor, torch.Tensor]] = None,
+) -> Union[torch.Tensor, Tuple[torch.Tensor, torch.Tensor]]:
+    """High-precision VFA API: handles quantization and the ``m_init`` estimate.
 
-    Mean-pools Q/K into per-block representatives, forms the block-level score
-    ``pooled_q @ pooled_k^T * d**-0.5``, softmaxes over key blocks, and keeps the
-    smallest set of top key blocks whose cumulative probability reaches
-    ``cdfthreshd`` per query block.  Blocks that fail the intra-block similarity
-    test are forced on (their mean is not a faithful summary, so they must be
-    evaluated exactly).
+    Thin convenience wrapper around :func:`fav3_sage_func`. It quantizes the
+    high-precision Q/K/V, builds the per-row frozen-max estimate ``m_init``, and
+    runs the (now unified) sage kernel with the VFA precomputed-max path.
 
-    Args:
-        attention_scored_only: when ``True``, skip the intra-block similarity
-            test and block-selection logic entirely, returning
-            ``(None, pooled_score)`` where ``pooled_score`` is the raw
-            block-level score ``pooled_q @ pooled_k^T * d**-0.5``.
+    The per-row frozen-max estimate ``m_init`` is the top-``n_sample_blocks``
+    blocks per q-block ranked by a SpargeAttn mean-pooled block-score and then
+    evaluated with real K rows (a lower bound on the true per-row max; no safety
+    margin). See :func:`compute_m_proxy_topn`.
 
-    Returns:
-        A ``(final_map, pooled_score)`` tuple.  ``final_map`` is a
-        ``(B, H, num_q_blocks, num_k_blocks)`` bool mask (or ``None`` when
-        ``attention_scored_only`` is set), and ``pooled_score`` is the
-        block-level score with positions that ``final_map`` masks out set to
-        ``-inf``.
+    When ``block_lut`` (a ragged ``(kv_block_indices, lut_start, lut_count)``
+    LUT from :func:`block_attn_mask_to_ragged_lut`) is provided, the
+    block-sparse path runs (the hot loop visits only attended K blocks); the
+    same guided ``m_init`` estimate is used.
     """
-    nq = (q.shape[-2] + BLKQ - 1) // BLKQ
-    nk = (k.shape[-2] + BLKK - 1) // BLKK
-    pooled_q, sim_q = get_pool_sim_triton_simmean(q, BLKQ, simthreshd1, attention_scored_only)
-    pooled_k, sim_k = get_pool_sim_triton_simmean(k, BLKK, simthreshd1, attention_scored_only)
-    pooled_score = pooled_q @ pooled_k.transpose(-1, -2) * q.shape[-1] ** -0.5
-    if attention_scored_only:
-        return None, pooled_score
+    assert q.dtype in [torch.float16, torch.bfloat16, torch.float32]
+    assert k.dtype in [torch.float16, torch.bfloat16, torch.float32]
+    assert v.dtype in [torch.float16, torch.bfloat16, torch.float32]
 
-    neg_inf = pooled_score.new_full((), float("-inf"))
-    sim_k = sim_k.unsqueeze(-2).expand(-1, -1, nq, -1)  # faster than repeat
-    sim_q = sim_q.unsqueeze(-1).expand(-1, -1, -1, nk)
+    if config is None:
+        config = get_sage_fwd_configs()
 
-    prob = torch.where(sim_k, pooled_score, neg_inf)
-    causal_mask = None
-    if is_causal:
-        causal_mask = fill_causal_mask_triton(
-            torch.empty(nq, nk, device=q.device, dtype=torch.bool), BLKQ / BLKK
-        )
-        prob = torch.where(causal_mask[None, None], prob, neg_inf)
-    prob = prob.softmax(-1)
+    bshd_map = [0, 1, 2, 3] if layout == "bshd" else [0, 2, 1, 3]
+    _, _, _, head_dim = map_dims(q.shape, bshd_map)
+    softmax_scale = softmax_scale or (head_dim ** -0.5)
 
-    # Keep the smallest set of top key blocks whose cumulative mass reaches cdfthreshd.
-    sorted_score = torch.sort(prob, dim=-1, descending=True)
-    cdf = sorted_score.values.cumsum(dim=-1)
-    H, K = cdf.shape[1], cdf.shape[-1]
-    ge = cdf >= cdfthreshd.view(1, H, 1, 1)
-    idx = ge.to(torch.uint8).argmax(dim=-1)
-    num_to_select = torch.where(ge.any(dim=-1), idx, idx.new_full((), K))
+    BLKQ, BLKK = config["BLOCK_M"], config["BLOCK_N"]
+    fp8_dtype = aiter.dtypes.fp8
+    fp8_max = torch.finfo(fp8_dtype).max
 
-    final_map = fill_block_map_triton(
-        torch.zeros_like(prob, dtype=torch.bool), num_to_select, sorted_score.indices
-    )
-    final_map = final_map | ~sim_k | ~sim_q
-    if is_causal:
-        final_map = final_map * causal_mask[None, None]
-    if attention_sink:
-        final_map[:, :, :, 0] = 1
-    return final_map, pooled_score
-
-
-def _num_text_blocks(text_len: int, block_m: int, block_n: int) -> Tuple[int, int]:
-    """Number of (q, k) blocks spanned by ``text_len`` trailing text tokens."""
-    return (
-        (text_len + block_m - 1) // block_m,
-        (text_len + block_n - 1) // block_n,
+    q_int8, q_descale, k_int8, k_descale, v_fp8, v_descale = sage_quant(
+        q, k, v,
+        fp8_dtype, fp8_max,
+        sm_scale=softmax_scale,
+        BLKQ=BLKQ,
+        BLKK=BLKK,
+        layout=layout,
     )
 
-
-def _assemble_full_block_mask(
-    image_block_mask: torch.Tensor,
-    image_len_q: int,
-    image_len_k: int,
-    text_len: int,
-    block_m: int,
-    block_n: int,
-) -> torch.Tensor:
-    """Append dense text rows/columns to an image-only block mask.
-
-    Returns the full ``(B, H, n_iq + n_text_q, n_ik + n_text_k)`` mask in which
-    all Q rows attend to the text K columns, all text Q rows attend to
-    everything, and any partial image/text boundary block is forced dense (so
-    spillover tokens are never dropped).  A no-op when ``text_len == 0``.
-    """
-    if text_len == 0:
-        return image_block_mask
-
-    B, H, n_iq, n_ik = image_block_mask.shape
-    n_text_q, n_text_k = _num_text_blocks(text_len, block_m, block_n)
-
-    full = torch.zeros(
-        (B, H, n_iq + n_text_q, n_ik + n_text_k),
-        dtype=image_block_mask.dtype,
-        device=image_block_mask.device,
-    )
-    full[:, :, :n_iq, :n_ik] = image_block_mask
-    full[:, :, :, -n_text_k:] = True  # every Q row attends to text K cols
-    full[:, :, -n_text_q:, :] = True  # text Q rows attend to everything
-    if image_len_q % block_m != 0:  # partial boundary blocks -> dense
-        full[:, :, image_len_q // block_m, :] = True
-    if image_len_k % block_n != 0:
-        full[:, :, :, image_len_k // block_n] = True
-    return full
-
-
-def block_attn_mask_to_ragged_lut_topn_front(
-    block_attn_mask: torch.Tensor,
-    pooled_score: torch.Tensor,
-    sample_n: int,
-    num_heads: Optional[int] = None,
-    force_front_mask: Optional[torch.Tensor] = None,
-) -> Tuple[torch.Tensor, torch.Tensor, torch.Tensor]:
-    """Build a ragged LUT with the top-``sample_n`` scored K blocks emitted first.
-
-    Like :func:`aiter.ops.triton.attention.utils.block_attn_mask_to_ragged_lut`,
-    this turns a block attention mask into a ragged
-    ``(kv_block_indices, lut_start, lut_count)`` LUT, but within each
-    ``(batch, head, q_block)`` segment it emits blocks in this order:
-
-      1. the ``sample_n`` highest ``pooled_score`` blocks, descending score;
-      2. the ``force_front_mask`` blocks (e.g. text), ascending block index;
-      3. the remaining attended blocks, ascending block index.
-
-    Building the LUT directly from the mask lets us write each segment out in the
-    desired order in a single pass -- no separate reorder of a pre-built LUT. The
-    valid entries are compactly packed into the prefix of ``kv_block_indices``
-    (indexed by ``lut_start``/``lut_count``); the buffer itself is over-allocated
-    to a static size so the whole function is CUDA-graph-safe (no data-dependent
-    shapes or device syncs), matching
-    :func:`aiter.ops.triton.attention.utils.block_attn_mask_to_ragged_lut`.
-
-    This pairs with the ``freeze_softmax_max_count`` block-sparse path
-    (:func:`fav3_sage_func`): the online-softmax running max is frozen after the
-    first few inner-loop iterations, so visiting the highest-scoring (and thus
-    likely highest-max) tiles -- plus any always-attended ``force_front`` tiles --
-    first makes the frozen max a tight estimate. See ``fav3_sage_vfa.py`` for the
-    analogous pooled-score top-N block selection.
-
-    Args:
-        block_attn_mask: ``(B, num_q_blocks, num_k_blocks)`` (shared across heads)
-            or ``(B, H, num_q_blocks, num_k_blocks)`` bool mask. True = attend.
-        pooled_score: fp32 ``(B, H, num_q_blocks, num_k_blocks)`` block-level
-            attention score (e.g. from :func:`get_block_map_meansim`). Only blocks
-            that are attended in ``block_attn_mask`` are ever selected.
-        sample_n: number of top-scored tiles to emit first per segment. ``<= 0``
-            emits only the ``force_front`` tiles ahead of the rest.
-        num_heads: number of Q heads; required when ``block_attn_mask`` is 3D.
-        force_front_mask: optional bool mask, same shape/broadcast as
-            ``block_attn_mask``, of blocks to place immediately after the sampled
-            tiles. Excluded from the top-``sample_n`` selection; only the
-            attended ones are emitted.
-
-    Returns:
-        ``kv_block_indices`` (1D int32), ``lut_start`` (1D int32) and
-        ``lut_count`` (1D int32), indexed by
-        ``idx = b * (H * num_q_blocks) + h * num_q_blocks + q_block``.
-        ``kv_block_indices`` is over-allocated to the static size ``B*H*Q*K``;
-        only the compact prefix selected by ``lut_start``/``lut_count`` is valid
-        (the remainder is unused padding).
-    """
-    if block_attn_mask.dim() == 3:
-        if num_heads is None:
-            raise ValueError("num_heads must be provided when block_attn_mask is 3D")
-        B, Q, K = block_attn_mask.shape
-        block_attn_mask = block_attn_mask.unsqueeze(1).expand(B, num_heads, Q, K)
-        if force_front_mask is not None and force_front_mask.dim() == 3:
-            force_front_mask = force_front_mask.unsqueeze(1).expand(B, num_heads, Q, K)
-
-    B, H, Q, K = block_attn_mask.shape
-    assert pooled_score.shape[:3] == (B, H, Q) and pooled_score.shape[-1] == K, (
-        f"pooled_score shape {tuple(pooled_score.shape)} does not match mask "
-        f"{(B, H, Q, K)}"
-    )
-    device = block_attn_mask.device
-
-    attended = block_attn_mask.to(torch.bool)
-    lut_count = attended.sum(-1).to(torch.int32).reshape(-1)
-    lut_start = torch.cumsum(lut_count, 0) - lut_count
-
-    if force_front_mask is None:
-        force_front = torch.zeros_like(attended)
+    use_block_sparse = block_lut is not None
+    if use_block_sparse:
+        kv_block_indices, lut_start, lut_count = block_lut
     else:
-        # Only attended blocks can be emitted at all.
-        force_front = force_front_mask.to(torch.bool).expand(B, H, Q, K) & attended
+        kv_block_indices = lut_start = lut_count = None
 
-    neg_inf = pooled_score.new_full((), float("-inf"))
-    masked_score = torch.where(attended, pooled_score.to(torch.float32), neg_inf)
-
-    # Mark the top-``sample_n`` attended, non-force-front blocks per (B, H, Q) row.
-    is_topn = torch.zeros((B, H, Q, K), dtype=torch.bool, device=device)
-    n = min(sample_n, K)
-    if n > 0:
-        sample_score = torch.where(force_front, neg_inf, masked_score)
-        topk = sample_score.topk(n, dim=-1)
-        # A row with fewer than n candidates pads topk with -inf entries; mark
-        # only the finite (genuinely attended, non-force-front) selections.
-        is_topn.scatter_(-1, topk.indices, topk.values > neg_inf)
-
-    # Per-row ordering of the K blocks by (priority, tiebreak):
-    #   0 = attended & top-n      -> descending score (highest-max first)
-    #   1 = attended & force-front -> ascending block index
-    #   2 = attended, the rest    -> ascending block index
-    #   3 = not attended          -> sorts past the per-row count, so dropped
-    col = torch.arange(K, device=device).view(1, 1, 1, K)
-    priority = torch.where(
-        ~attended,
-        3,
-        torch.where(is_topn, 0, torch.where(force_front, 1, 2)),
-    )
-    tiebreak = torch.where(is_topn, -masked_score, col.to(torch.float32))
-
-    # Lexicographic (priority, tiebreak) sort per row via two stable sorts.
-    o1 = torch.argsort(tiebreak, dim=-1, stable=True)
-    order = torch.gather(
-        o1, -1, torch.argsort(torch.gather(priority, -1, o1), dim=-1, stable=True)
+    m_init = compute_m_proxy_topn(
+        q, k, q_int8, k_int8, q_descale, k_descale,
+        BLKQ=BLKQ, BLKK=BLKK, layout=layout,
+        n_blocks=n_sample_blocks,
     )
 
-    # The first ``count`` entries of each row are exactly the attended blocks in
-    # the desired order; pack them row-major into the ragged index list. Done as
-    # a scatter into an over-allocated, statically-sized buffer (rather than
-    # boolean-mask indexing, which triggers a ``nonzero`` device sync and a
-    # data-dependent output shape) so the function stays CUDA-graph-safe.
-    R = B * H * Q
-    rows = order.reshape(R, K)
-    col = torch.arange(K, device=device)
-    keep = col[None, :] < lut_count[:, None]
-    # Destination of each entry in the packed buffer. Dropped (unattended)
-    # entries are routed to a scratch sink slot so they can never clobber a
-    # valid neighbouring row's position.
-    sink = R * K
-    dest = torch.where(
-        keep,
-        lut_start[:, None].to(torch.long) + col[None, :],
-        col.new_full((), sink),
+    out, lse = fav3_sage_func(
+        q_int8, k_int8, v_fp8,
+        q_descale, k_descale, v_descale,
+        softmax_scale=softmax_scale,
+        return_lse=return_lse,
+        layout=layout,
+        config=config,
+        kv_block_indices=kv_block_indices,
+        lut_start=lut_start,
+        lut_count=lut_count,
+        use_block_sparse=use_block_sparse,
+        m_init=m_init,
     )
-    packed = torch.empty(sink + 1, dtype=torch.int32, device=device)
-    packed.scatter_(0, dest.reshape(-1), rows.reshape(-1).to(torch.int32))
-    kv_block_indices = packed[:sink]
-    return kv_block_indices, lut_start, lut_count
-
-
-def build_attention_lut(
-    q: torch.Tensor,
-    k: torch.Tensor,
-    *,
-    simthreshd1: float,
-    cdfthreshd: float,
-    mode: str = "both",
-    n_sample: int = 8,
-    is_causal: bool = False,
-    static_block_mask: Optional[torch.Tensor] = None,
-    text_len: int = 0,
-    block_m: int = 128,
-    block_n: int = 128,
-) -> Tuple[Tuple[torch.Tensor, torch.Tensor, torch.Tensor], int]:
-    """Build a block-sparse ragged LUT for SpargeAttn and/or VFA.
-
-    One entry point that turns Q/K into a ready-to-use ragged LUT plus the
-    matching ``freeze_softmax_max_count``. ``mode`` selects the block set and
-    ordering:
-
-      * ``"sparge"``: keep only the SpargeAttn-selected (meansim) blocks and emit
-        them in ascending order. ``freeze_softmax_max_count = -1`` (the online
-        softmax max is never frozen); ``n_sample`` is ignored.
-      * ``"vfa"``: keep *all* K blocks (block-causal when ``is_causal``) but
-        front-load the top-``n_sample`` blocks by pooled score, so the frozen max
-        sees the most important tiles first.
-      * ``"both"``: keep only the SpargeAttn-selected blocks *and* front-load the
-        top-``n_sample`` of them by pooled score.
-
-    Text handling (``text_len > 0``): every q-block additionally attends to the
-    trailing dense text K blocks. For ``"vfa"``/``"both"`` those text blocks are
-    appended to the front immediately *after* the ``n_sample`` sampled tiles, and
-    the returned ``freeze_softmax_max_count`` is ``n_sample + n_text_blocks`` so
-    the max is frozen only once both the sampled and text tiles are visited.
-
-    Causality is handled at block granularity (whole K blocks are kept/dropped);
-    the block-sparse kernel does not apply an intra-block diagonal mask.
-
-    Args mirror :func:`compute_sparge_block_mask`, plus ``mode`` and ``n_sample``.
-
-    Returns:
-        ``(block_lut, freeze_softmax_max_count)`` where ``block_lut`` is the
-        ragged ``(kv_block_indices, lut_start, lut_count)`` tuple to pass as
-        ``block_lut`` to :func:`fav3_sage_wrapper_func` (or
-        ``kv_block_indices``/``lut_start``/``lut_count`` to :func:`fav3_sage_func`
-        with ``use_block_sparse=True`` and ``freeze_softmax_max_count=...``).
-    """
-    if mode not in ("vfa", "sparge", "both"):
-        raise ValueError(f"mode must be 'vfa', 'sparge' or 'both', got {mode!r}")
-
-    image_q = q[:, :, : q.shape[2] - text_len, :] if text_len > 0 else q
-    image_k = k[:, :, : k.shape[2] - text_len, :] if text_len > 0 else k
-    image_len_q = q.shape[2] - text_len
-    image_len_k = k.shape[2] - text_len
-    n_text_k = _num_text_blocks(text_len, block_m, block_n)[1] if text_len > 0 else 0
-
-    # "vfa" ranks every block (no sparsity selection), so it only needs the
-    # pooled score; "sparge"/"both" also need the meansim sparsity mask.
-    image_mask, image_score = get_block_map_meansim(
-        image_q,
-        image_k,
-        is_causal=is_causal,
-        BLKQ=block_m,
-        BLKK=block_n,
-        simthreshd1=simthreshd1,
-        cdfthreshd=cdfthreshd,
-        attention_scored_only=(mode == "vfa"),
-    )
-
-    if mode == "vfa":
-        # Keep all blocks (block-causal when requested) as sampling candidates.
-        B, H, n_iq, n_ik = image_score.shape
-        if is_causal:
-            causal = fill_causal_mask_triton(
-                torch.empty(n_iq, n_ik, device=q.device, dtype=torch.bool),
-                block_m / block_n,
-            )
-            image_mask = causal[None, None].expand(B, H, n_iq, n_ik).clone()
-        else:
-            image_mask = torch.ones((B, H, n_iq, n_ik), dtype=torch.bool, device=q.device)
-
-    if static_block_mask is not None:
-        image_mask = image_mask | static_block_mask[None, None, ...]
-
-    full_mask = _assemble_full_block_mask(
-        image_mask, image_len_q, image_len_k, text_len, block_m, block_n
-    )
-
-    if mode == "sparge":
-        return block_attn_mask_to_ragged_lut(full_mask), -1
-
-    # vfa / both: front-load the top-n sampled image tiles, then the dense text
-    # tiles. Scores live only over the image region; text/text-row positions stay
-    # -inf so they are never picked as sampled tiles (text is forced front).
-    B, H, n_iq, n_ik = image_mask.shape
-    n_tq, n_tk = full_mask.shape[-2], full_mask.shape[-1]
-    full_score = full_mask.new_full(
-        (B, H, n_tq, n_tk), float("-inf"), dtype=torch.float32
-    )
-    full_score[:, :, :n_iq, :n_ik] = image_score.to(torch.float32)
-
-    force_front = None
-    if n_text_k > 0:
-        force_front = torch.zeros((B, H, n_tq, n_tk), dtype=torch.bool, device=q.device)
-        force_front[:, :, :, -n_text_k:] = True
-
-    block_lut = block_attn_mask_to_ragged_lut_topn_front(
-        full_mask, full_score, n_sample, force_front_mask=force_front
-    )
-    return block_lut, n_sample + n_text_k
+    if return_lse:
+        return out, lse
+    return out
